Rename package WikipediaExternal->WikipediaData
