+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "libs/WikipediaExternal"]
-	path = libs/WikipediaExternal
+[submodule "libs/WikipediaData"]
+	path = libs/WikipediaData
 	url = https://github.com/AldanaNonnenmacher/AyDS23-Winchester2-SubModuleWikipedia.git
