+++ app/src/test/java/ayds/winchester/songinfo/moredetails/presentation/InfoDescriptionHelperTest.kt
@@ -1,6 +1,6 @@
 package ayds.winchester.songinfo.moredetails.presentation
 
-import ayds.winchester2.wikipediaexternal.data.wikipedia.entity.Info.ArtistInfo
+import ayds.winchester2.wikipediadata.data.wikipedia.entity.Info.ArtistInfo
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Assert.assertEquals
