+++ app/src/test/java/ayds/winchester/songinfo/moredetails/presentation/MoreDetailsPresenterTest.kt
@@ -1,7 +1,7 @@
 package ayds.winchester.songinfo.moredetails.presentation
 
-import ayds.winchester2.wikipediaexternal.data.wikipedia.entity.Info.ArtistInfo
-import ayds.winchester2.wikipediaexternal.data.wikipedia.entity.Info.EmptyInfo
+import ayds.winchester2.wikipediadata.data.wikipedia.entity.Info.ArtistInfo
+import ayds.winchester2.wikipediadata.data.wikipedia.entity.Info.EmptyInfo
 import ayds.winchester.songinfo.moredetails.domain.repository.WikipediaRepository
 import io.mockk.every
 import io.mockk.mockk
