+++ settings.gradle
@@ -1,5 +1,5 @@
 include ':app'
 include 'observer'
 rootProject.name = "AK Winchester"
-include ':wikipediaexternal'
-project(':wikipediaexternal').projectDir = new File('libs/WikipediaExternal')
+include ':wikipediadata'
+project(':wikipediadata').projectDir = new File('libs/WikipediaData')
